requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
click==8.1.3
Django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
